Added shape versatility
Added a initialXYcurve struct that basically stores the XY curve of the 3d shape we want to draw. It's eqation and it's theta limits. like for a torus it is {R2 + R1 cos theta, R1 sin theta, 0} and for a cylinder it will be something like {r, (h/2) sin theta, 0} and the corresponding limits will be (0, 2pi) for torus and (-pi/2 , pi/2) for cylinder